requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
mysql-connector-python==8.2.0
bcrypt==4.1.2
